+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 2.3.13 - UNRELEASED
 
 * Update glTF schema to latest (document revision 2.0.1).
-* Added JSON schema for KHR_xmp_json_ld.
+* Added JSON schema for KHR_materials_emissive_strength, KHR_materials_iridescence, and KHR_xmp_json_ld.
 
 ### 2.3.12 - 2021-08-03