+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ### 2.3.13 - UNRELEASED
 
-* Update glTF schema to latest (document revision 2.0.1).
-* Added JSON schema for KHR_materials_emissive_strength, KHR_materials_iridescence, and KHR_xmp_json_ld.
-* Added the first "Quick Fix": Add an undeclared extension to `extensionsUsed`.
+* Update glTF schema to latest (document revision 2.0.1). [#226](https://github.com/AnalyticalGraphicsInc/gltf-vscode/pull/226)
+* Added JSON schema for KHR_materials_emissive_strength, KHR_materials_iridescence, and KHR_xmp_json_ld. Included in [#226](https://github.com/AnalyticalGraphicsInc/gltf-vscode/pull/226)
+* Added the first "Quick Fix": Add an undeclared extension to `extensionsUsed`. [#227](https://github.com/AnalyticalGraphicsInc/gltf-vscode/pull/227)
 
 ### 2.3.12 - 2021-08-03